#OD-17 Lots of refactoring to the DIsplayResultls Move elapsed time to Shared Add Club Images

+++ CourseVariant.cs
@@ -7,7 +7,7 @@ namespace OR_Results
         public int CourseVariantId { get; set; }
         public string CourseId { get; set; }
         public string Field1 { get; set; }
-        public int Field2 { get; set; }
+        public double Field2 { get; set; }
         public string Field3 { get; set; }
         public int Field4 { get; set; }
         public int Field5 { get; set; }
